+++ src/main/java/com/mysql/MySqlFilmshowDao.java
@@ -12,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManagerFactory;
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -107,7 +108,7 @@ public class MySqlFilmshowDao implements FilmshowDao {
     }
 
     @Override
-    public Filmshow getFilmshowById(int id) throws DaoException {
+    public Filmshow getFilmshowById(BigInteger id) throws DaoException {
         try {
             session = sessionFactory.openSession();
             session.beginTransaction();
