+++ src/main/java/com/web/FilmshowEditor.java
@@ -4,6 +4,7 @@ import com.domain.Filmshow;
 import org.springframework.stereotype.Component;
 
 import java.beans.PropertyEditorSupport;
+import java.math.BigInteger;
 
 @Component
 public class FilmshowEditor extends PropertyEditorSupport {
@@ -12,7 +13,7 @@ public class FilmshowEditor extends PropertyEditorSupport {
     public void setAsText(String text) {
         try {
             Filmshow filmshow = new Filmshow();
-            filmshow.setFilmshowId(Integer.parseInt(text));
+            filmshow.setFilmshowId(BigInteger.valueOf(Integer.parseInt(text)));
             this.setValue(filmshow);
         } catch (Exception e) {
             e.printStackTrace();
